// БР1.2/Персук Виктория/labs/lab2/services/restaurant-service/src/common/enums.ts
export enum RoleName {
    Admin = 'Admin',
    User = 'User',
    Owner = 'Owner',
    Manager = 'Manager',
}

export enum RestaurantStatus {
    Pending = 'Pending',
    Verified = 'Verified',
    Rejected = 'Rejected',
}

export enum PriceCategory {
    Low = '$',
    Medium = '$$',
    High = '$$$',
}